registro de clinicas y usuarios
