+++ src/pages/register/register-clinic/index.js
@@ -1,0 +1,1 @@
+export * from './register-clinic-page';
