+++ src/pages/register/register-user/index.js
@@ -1,0 +1,1 @@
+export * from './register-user-page';
